+++ src/main/java/com/hims/request/LabInvestigationReq.java
@@ -3,7 +3,6 @@ package com.hims.request;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Date;
 
@@ -13,8 +12,8 @@ public class LabInvestigationReq {
     private Long id;
     private LocalDate appointmentDate;
    // private boolean checkStatus;
-    private BigDecimal actualAmount;
-    private BigDecimal discountedAmount;
+    private int actualAmount;
+    private int discountedAmount;
     private String type;
